shoeshop_project/requirements/prod-requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
bleach==6.1.0
bx_py_utils==91
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-countries==7.5.1
django-debug-toolbar==4.3.0
django-dynamic-breadcrumbs==0.4.3
django-redis==5.4.0
django-timezone-field==6.1.0
django-tools==0.54.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==23.2.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
icdiff==2.0.7
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
Pillow==10.0.0
pluggy==1.4.0
pprintpp==0.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.4
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
